requirements.txt
backports.cached-property==1.0.1
grpcio==1.43.0
grpcio-tools==1.43.0
overrides==6.1.0
protobuf==3.19.1
robotframework==4.1.3
robotframework-assertion-engine==0.3.1
robotframework-browser==11.1.1
robotframework-pythonlibcore==3.0.0
six==1.16.0
typing-extensions==4.0.1
typing-utils==0.1.0
wrapt==1.13.3
